feat: positions table overhaul — merged wheel, strategy badges, smart notes, flag tooltips, qty/avg columns
- Merge WheelSP/WheelSC positions into the main table (41 total)
- Strategy column now shows colored bordered badges (LEAP=indigo, LDS=violet, SWING=yellow, 2x-ETF=orange, Thematic=cyan, WheelSP=emerald, WheelSC=teal)
- Per-strategy note normalization: strips theme/strategy prefix, surfaces movement %, duration, Fwd R/R, strike/expiry for each strategy type
- Flag dots now show a fixed-position styled tooltip on hover (escapes table overflow clipping)
- Add Qty column (contracts/shares) and Avg column (cost basis per unit)
- IBKR bridge agent + SSE refresh stream + system API endpoint
- Sidebar stats panel with live NAV/daily P&L; status bar with clock and refresh button
- DB helpers get proper commit/rollback; api.ts fixed for server-side absolute URL

+++ backend/app/api/analytics.py
@@ -14,9 +14,9 @@ def scorecard():
                 100.0 * COUNT(*) FILTER (WHERE realized_pnl > 0 AND close_date >= CURRENT_DATE - INTERVAL '1 year')
                 / NULLIF(COUNT(*) FILTER (WHERE close_date >= CURRENT_DATE - INTERVAL '1 year'), 0)
             , 0)                                                                            AS win_rate,
-            ROUND(AVG(gain_pct) FILTER (WHERE close_date >= CURRENT_DATE - INTERVAL '1 year')::numeric, 1) AS avg_gain_pct,
-            ROUND(SUM(realized_pnl) FILTER (WHERE EXTRACT(YEAR FROM close_date) = EXTRACT(YEAR FROM CURRENT_DATE))::numeric, 0) AS ytd_pnl,
-            ROUND(SUM(realized_pnl) FILTER (WHERE close_date >= CURRENT_DATE - INTERVAL '1 year')::numeric, 0) AS pnl_1y
+            ROUND(AVG(gain_pct) FILTER (WHERE close_date >= CURRENT_DATE - INTERVAL '1 year'), 1) AS avg_gain_pct,
+            ROUND(SUM(realized_pnl) FILTER (WHERE EXTRACT(YEAR FROM close_date) = EXTRACT(YEAR FROM CURRENT_DATE)), 0) AS ytd_pnl,
+            ROUND(SUM(realized_pnl) FILTER (WHERE close_date >= CURRENT_DATE - INTERVAL '1 year'), 0) AS pnl_1y
         FROM closed_trades
         WHERE account_id = 1 AND strategy IS NOT NULL
         GROUP BY strategy
@@ -46,9 +46,9 @@ def by_strategy():
     return query("""
         SELECT
             strategy,
-            ROUND(SUM(realized_pnl) FILTER (WHERE EXTRACT(YEAR FROM close_date) = EXTRACT(YEAR FROM CURRENT_DATE))::numeric, 0) AS ytd,
-            ROUND(SUM(realized_pnl) FILTER (WHERE close_date >= CURRENT_DATE - INTERVAL '1 year')::numeric, 0)                  AS one_year,
-            COUNT(*) FILTER (WHERE EXTRACT(YEAR FROM close_date) = EXTRACT(YEAR FROM CURRENT_DATE))                             AS ytd_trades
+            ROUND(SUM(realized_pnl) FILTER (WHERE EXTRACT(YEAR FROM close_date) = EXTRACT(YEAR FROM CURRENT_DATE)), 0) AS ytd,
+            ROUND(SUM(realized_pnl) FILTER (WHERE close_date >= CURRENT_DATE - INTERVAL '1 year'), 0)                  AS one_year,
+            COUNT(*) FILTER (WHERE EXTRACT(YEAR FROM close_date) = EXTRACT(YEAR FROM CURRENT_DATE))                    AS ytd_trades
         FROM closed_trades
         WHERE account_id = 1 AND strategy IS NOT NULL
         GROUP BY strategy
